ml_service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn>=1.5.0
numpy>=1.24.3
pandas>=2.0.3
python-dotenv==1.0.0